# Conflicts:
#	doc/03模块设计/外部接口_我的_我是招标人.md
#	src/server/Paas/.classpath
#	src/server/Paas/.gitignore
#	src/server/Paas/.project
#	src/server/Paas/src/main/java/com/hummingbird/paas/controller/CapitalManageController.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/controller/MyBidderBusinessController.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/controller/UserCenterController.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/controller/UserSecurityController.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/entity/BidRecord.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/entity/BiddeeCertificateAduit.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/entity/BidderCertificateAduit.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/entity/Order.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/entity/ProjectAccount.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/entity/ProjectEvaluationBiddee.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/entity/ProjectEvaluationBidder.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/entity/ProjectPaymentDefineDetail.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/exception/PaasException.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/BidRecordMapper.xml
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/BiddeeCerticateMapper.xml
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/BiddeeCertificateAduitMapper.xml
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/BiddeeCreditMapper.xml
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/BiddeeMapper.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/BiddeeMapper.xml
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/BidderCerticateMapper.xml
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/BidderCertificateAduitMapper.xml
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/BidderCertificationMapper.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/OrderMapper.xml
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/ProjectAccountOrderMapper.xml
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/ProjectEvaluationBiddeeMapper.xml
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/ProjectPaymentDefineDetailMapper.xml
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/ProjectPaymentPayMapper.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/ProjectPaymentPayMapper.xml
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/ProjectPaymentReceiveMapper.xml
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/ProjectPaymentRecordMapper.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/ProjectPaymentRecordMapper.xml
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/UserBankcardMapper.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/mapper/UserBankcardMapper.xml
#	src/server/Paas/src/main/java/com/hummingbird/paas/services/MyBidderService.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/services/impl/MyBidderServiceImpl.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/services/impl/ProjectServiceImpl.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/services/impl/UserServiceImpl.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/util/MoneyUtil.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/BuyTenderMemberBodyVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/ForgetPasswordBodyVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/ForgetPasswordVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/FreezeBondBodyVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/FreezeBondReturnVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/FreezeWithdrawalsBodyVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/LoginBodyVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/LoginVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/RechargeApplyBodyVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/UnfreezeVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/UpdateLoginPasswordBodyVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/UpdateLoginPasswordVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/UpdateMobileNumBodyVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/UpdateMobileNumVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/UpdateTradePasswordBodyVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/UpdateTradePasswordVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/UserSecurityDetailVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/UserSecurityInfoVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/VerifySmsCodeVO.java
#	src/server/Paas/src/main/java/com/hummingbird/paas/vo/WithdrawalsApplyBodyVO.java
#	src/server/capital/target/classes/.gitignore
#	src/server/capital/target/classes/applicationContext.xml
#	src/server/capital/target/classes/config/dev/jdbc.properties
#	src/server/capital/target/classes/config/dev/log4j.properties
#	src/server/capital/target/classes/config/pro/common.properties
#	src/server/capital/target/classes/config/pro/jdbc.properties
#	src/server/capital/target/classes/config/pro/log4j.properties
#	src/server/capital/target/classes/config/sit/common.properties
#	src/server/capital/target/classes/config/sit/jdbc.properties
#	src/server/capital/target/classes/config/sit/log4j.properties
#	src/server/capital/target/classes/config/uat/common.properties
#	src/server/capital/target/classes/config/uat/jdbc.properties
#	src/server/capital/target/classes/config/uat/log4j.properties
#	src/server/capital/target/classes/dataSource.xml
#	src/server/capital/target/classes/jdbc.properties
#	src/server/capital/target/classes/log4j.properties
#	src/server/capital/target/classes/spring-scheduler.xml
#	src/server/capital/target/classes/spring-servlet.xml
#	src/server/capital/target/m2e-wtp/web-resources/META-INF/maven/com.hummingbird/capital/pom.xml
#	src/server/capital/target/m2e-wtp/web-resources/META-INF/maven/com.hummingbird/paas/pom.xml
#	src/server/portal/Application/Admin/Controller/ZtglobjectController.class.php
#	src/server/portal/Application/Admin/View/ace/Ztglobject/viewzhongbiao.php
#	src/server/portal/Root/Public/Admin/css/common.css
#	src/server/userCenter/.classpath
#	src/server/userCenter/.gitignore

+++ src/server/Paas/src/main/java/com/hummingbird/paas/entity/MakeMatchBondRecord.java
@@ -24,7 +24,7 @@ public class MakeMatchBondRecord {
     /**
      * 金额，单位为分
      */
-    private Integer bondAmount;
+    private Long bondAmount;
 
     /**
      * 状态，FOZ冻结（帐户内的钱），PAY缴（保证金，中标后），REV收保证金
@@ -99,7 +99,7 @@ public class MakeMatchBondRecord {
     /**
      * @return 金额，单位为分
      */
-    public Integer getBondAmount() {
+    public Long getBondAmount() {
         return bondAmount;
     }
 
@@ -107,7 +107,7 @@ public class MakeMatchBondRecord {
      * @param bondAmount 
 	 *            金额，单位为分
      */
-    public void setBondAmount(Integer bondAmount) {
+    public void setBondAmount(Long bondAmount) {
         this.bondAmount = bondAmount;
     }
